Tests refactor
