+++ platformio.ini
@@ -10,13 +10,11 @@
 
 [env:native]
 platform = native
-test_filter = test_native
 
 [env:nano33ble]
 platform = nordicnrf52@9.5.0
 board = nano33ble
 framework = arduino
-test_filter = test_embedded
 ; Use arduino framework Unity version instead of Platformio default version
 ; https://docs.platformio.org/en/stable/advanced/unit-testing/frameworks/custom/examples/custom_unity_library.html
 test_framework = custom
@@ -26,7 +24,6 @@ platform = nordicnrf52@9.5.0
 board = nano33ble
 framework = arduino
 build_flags = -DRENODE_ENVIRONMENT
-test_filter = test_embedded
 ; Use arduino framework Unity version instead of Platformio default version
 ; https://docs.platformio.org/en/stable/advanced/unit-testing/frameworks/custom/examples/custom_unity_library.html
 test_framework = custom
